projects/oauth/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-jose
python-multipart
pydantic-settings